Proof of Concept for retrieving valid cover images
This branch is not intended to ever be merged/deployed, it exists as an exploration and proof-of-concept for fetching cover images through the GITenberg project and the main Project Gutenberg server. It grabs the files and stores them locally.

Importantly it only gets `archival` covers, or those from the original volumes. This is distinguished from `generated` covers that feature algorithmically generated patterns and which are not useful/interesting for our users.

The actual implementation of this work will involve storing the cover files in s3 and returning references to the ingest process. These can be handled as other updates in the sfr-db-updater and should create a link between the instance and the cover image.